restaurant_system/requirements.txt
# requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
pika>=1.3.2